@@ -1,11 +1,12 @@
 {
   "name": "coinstac-ui",
-  "version": "4.7.5",
+  "version": "5.0.3",
   "description": "COINSTAC desktop application.",
   "main": "app/main/index.js",
   "dependencies": {
-    "@material-ui/core": "^3.6.1",
-    "@material-ui/icons": "^3.0.1",
+    "@material-ui/core": "^4.0.0",
+    "@material-ui/icons": "^4.0.0",
+    "@material-ui/styles": "^4.0.0",
     "add-graphql-subscriptions": "^1.0.1",
     "app-module-path": "^1.1.0",
     "async": "^2.6.1",
@@ -15,11 +16,11 @@
     "bootstrap-sass": "^3.3.7",
     "clarify": "^2.1.0",
     "classnames": "^2.2.5",
-    "coinstac-client-core": "^4.7.5",
-    "coinstac-common": "^4.7.4",
-    "coinstac-docker-manager": "^4.7.4",
-    "coinstac-graphql-schema": "^4.7.1",
-    "coinstac-pipeline": "^4.7.5",
+    "coinstac-client-core": "^5.0.3",
+    "coinstac-common": "^5.0.3",
+    "coinstac-docker-manager": "^5.0.3",
+    "coinstac-graphql-schema": "^5.0.0",
+    "coinstac-pipeline": "^5.0.3",
     "commander": "^2.15.1",
     "convict": "^1.5.0",
     "create-react-class": "^15.6.3",
@@ -41,14 +42,14 @@
     "notistack": "^0.9.9",
     "prop-types": "^15.6.1",
     "rasterizehtml": "^1.3.0",
-    "react": "^16.5.2",
+    "react": "^16.8.0",
     "react-apollo": "^1.4.16",
     "react-avatar": "^2.5.1",
     "react-bootstrap": "^0.31.5",
     "react-bootstrap-typeahead": "^3.1.4",
     "react-dnd": "^2.6.0",
     "react-dnd-html5-backend": "^2.6.0",
-    "react-dom": "^16.5.2",
+    "react-dom": "^16.8.0",
     "react-dragula": "^1.1.17",
     "react-faux-dom": "^4.1.0",
     "react-material-ui-form-validator": "^2.0.9",
@@ -115,6 +116,10 @@
     "webpack": "^4.20.2",
     "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.1.9"
+  },
+  "peerDependencies": {
+    "react": "^16.8.0",
+    "react-dom": "^16.8.0"
   },
   "scripts": {
     "build:webpack": "cross-env NODE_ENV=production webpack --mode=production",
