@@ -1,156 +1,154 @@
-const rethink = require('rethinkdb');
+/* eslint-disable no-await-in-loop */
 const Boom = require('boom');
 const GraphQLJSON = require('graphql-type-json');
 const Promise = require('bluebird');
 const { PubSub, withFilter } = require('graphql-subscriptions');
 const axios = require('axios');
 const { uniq } = require('lodash');
+const { ObjectID } = require('mongodb');
 const helperFunctions = require('../auth-helpers');
 const initSubscriptions = require('./subscriptions');
-const config = require('../../config/default');
+const database = require('../database');
+const { transformToClient } = require('../utils');
+const {
+  eventEmitter,
+  COMPUTATION_CHANGED,
+  COMPUTATION_DELETED,
+  CONSORTIUM_CHANGED,
+  CONSORTIUM_DELETED,
+  PIPELINE_CHANGED,
+  PIPELINE_DELETED,
+  RUN_CHANGED,
+  THREAD_CHANGED,
+  USER_CHANGED,
+} = require('./events');
 
-/**
- * Helper function to retrieve all members of given table
- * @param {string} table - The table name
- * @return {array} The contents of the requested table
- */
-function fetchAll(table) {
-  let connection;
-  return helperFunctions.getRethinkConnection()
-    .then((db) => {
-      connection = db;
-      return rethink.table(table).orderBy({ index: 'id' }).run(connection);
-    })
-    .then(cursor => cursor.toArray())
-    .then(res => connection.close().then(() => res));
-}
+async function fetchOnePipeline(id) {
+  const db = database.getDbInstance();
 
-/**
- * Helper function to retrieve a single entry in a table
- * @param {string} table - The table name
- * @param {string} id - The entry id
- * @return {object} The requested table entry
- */
-function fetchOne(table, id) {
-  return helperFunctions.getRethinkConnection()
-    .then(connection => rethink.table(table).get(id).run(connection));
-}
+  const pipelineSteps = await db.collection('pipelines').aggregate([
+    { $match: { _id: id } },
+    { $unwind: '$steps' },
+    {
+      $lookup: {
+        from: 'computations',
+        localField: 'steps.computations',
+        foreignField: '_id',
+        as: 'steps.computations',
+      },
+    },
+  ]);
 
-function fetchOnePipeline(id) {
-  return helperFunctions.getRethinkConnection()
-    .then(connection => rethink.table('pipelines')
-      .get(id)
-    // Populate computations subfield with computation meta information
-      .merge(pipeline => ({
-        steps: pipeline('steps').map(step => step.merge({
-          computations: step('computations').map(compId => rethink.table('computations').get(compId)),
-        })),
-      }))
-      .run(connection)
-      .then(res => connection.close().then(() => res)))
-    .then(result => result);
+  const hasSteps = await pipelineSteps.hasNext();
+  if (!hasSteps) {
+    return db.collection('pipelines').findOne({ _id: id });
+  }
+
+  let pipe = null;
+  do {
+    const currentStep = await pipelineSteps.next();
+
+    if (!pipe) {
+      pipe = {
+        ...currentStep,
+        steps: [],
+      };
+    }
+
+    currentStep.steps.computations = transformToClient(currentStep.steps.computations);
+
+    pipe.steps.push(currentStep.steps);
+  } while (await pipelineSteps.hasNext());
+
+  return pipe;
 }
 
 /**
  * Helper function for add permissions to an user
- * @param {object} connection - Existing db connection
  * @param {object} args - Update object
  * @param {string} args.userId - Id of the user which will have permissions changed
  * @param {string} args.role - Role of the user
  * @param {string} args.doc - Id of the document for which the user will gain access
  * @param {string} args.table - Table of the document for which the user will gain access
  */
-async function addUserPermissions(connection, args) {
-  const perms = await rethink.table('users').get(args.userId)('permissions').run(connection);
+async function addUserPermissions(args) {
+  const db = database.getDbInstance();
 
   const { role, doc, table } = args;
 
-  let newRoles = [role];
-  const promises = [];
+  const updateObj = {
+    $addToSet: {
+      [`permissions.${table}.${doc}`]: role,
+    },
+  };
 
-  // Grab existing roles if present
-  if (perms[table][doc] && perms[table][doc].indexOf(role) === -1) {
-    newRoles = newRoles.concat(perms[table][doc]);
-  } else if (perms[table][doc]) {
-    newRoles = perms[table][doc];
-  }
-
-  const updateObj = { permissions: { [table]: { [doc]: newRoles } } };
-
-  // Add entry to user statuses object &&
   if (table === 'consortia') {
-    updateObj.consortiaStatuses = {};
-    updateObj.consortiaStatuses[doc] = 'none';
+    updateObj.$set = {
+      [`consortiaStatuses.${doc}`]: 'none',
+    };
 
-    promises.push(
-      rethink.table('consortia').get(doc).update(
-        {
-          [`${role}s`]: rethink.row(`${role}s`).append(args.userId),
-        }
-      ).run(connection)
-    );
+    const consortiaUpdateResult = await db.collection('consortia').findOneAndUpdate({ _id: doc }, {
+      $addToSet: { [`${role}s`]: args.userId },
+    }, { returnOriginal: false });
+
+    eventEmitter.emit(CONSORTIUM_CHANGED, consortiaUpdateResult.value);
   }
 
-  promises.push(
-    rethink.table('users').get(args.userId).update(
-      updateObj, { returnChanges: true }
-    ).run(connection)
-  );
+  const userUpdateResult = await db.collection('users').findOneAndUpdate({ _id: args.userId }, updateObj, { returnOriginal: false });
 
-  return Promise.all(promises);
+  eventEmitter.emit(USER_CHANGED, userUpdateResult.value);
 }
 
-async function removeUserPermissions(connection, args) {
-  const promises = [];
+async function removeUserPermissions(args) {
+  const db = database.getDbInstance();
 
-  const nextPermissions = await rethink.table('users')
-    .get(args.userId)('permissions')(args.table)(args.doc)
-    .filter(role => role.ne(args.role))
-    .run(connection);
+  const user = await db.collection('users').findOne({ _id: args.userId }, {
+    projection: { permissions: 1 },
+  });
 
-  if (nextPermissions.length === 0) {
-    const replaceObj = {
-      permissions: { [args.table]: args.doc },
+  const { permissions } = user;
+
+  const index = permissions[args.table][args.doc].findIndex(p => p === args.role);
+  permissions[args.table][args.doc].splice(index, 1);
+
+  let userUpdateResult;
+
+  if (permissions[args.table][args.doc].length === 0) {
+    const updateObj = {
+      $unset: {
+        [`permissions.${args.table}.${args.doc}`]: '',
+      },
     };
 
     if (args.table === 'consortia') {
-      replaceObj.consortiaStatuses = args.doc;
-
-      await rethink.table('consortia')
-        .get(args.doc)
-        .update({ mappedForRun: rethink.row('mappedForRun').difference([args.userId]) })
-        .run(connection);
+      updateObj.$unset[`consortiaStatuses.${args.doc}`] = '';
     }
 
-    promises.push(
-      rethink.table('users').get(args.userId).replace(user => user.without(replaceObj), { nonAtomic: true }).run(connection)
-    );
+    userUpdateResult = await db.collection('users').findOneAndUpdate({ _id: args.userId }, updateObj, { returnOriginal: false });
   } else {
-    promises.push(
-      rethink.table('users')
-        .get(args.userId)
-        .update({
-          permissions: {
-            [args.table]: {
-              [args.doc]: rethink.row('permissions')(args.table)(args.doc).difference([args.role]),
-            },
-          },
-        }, { nonAtomic: true })
-        .run(connection)
-    );
+    userUpdateResult = await db.collection('users').findOneAndUpdate({ _id: args.userId }, {
+      $pull: { [`permissions.${args.table}.${args.doc}`]: args.role },
+    }, {
+      returnOriginal: false,
+    });
   }
+
+  eventEmitter.emit(USER_CHANGED, userUpdateResult.value);
 
   if (args.table === 'consortia') {
-    promises.push(
-      rethink.table('consortia').get(args.doc).update(
-        {
-          [`${args.role}s`]: rethink.row(`${args.role}s`).difference([args.userId]),
-        }
-      ).run(connection)
-    );
-  }
+    const updateObj = {
+      $pull: {
+        [`${args.role}s`]: args.userId,
+      },
+    };
 
-  await Promise.all(promises);
+    if (permissions[args.table][args.doc].length === 0) {
+      updateObj.$pull.mappedForRun = args.userId;
+    }
+
+    const consortiaUpdateResult = await db.collection('consortia').findOneAndUpdate({ _id: args.doc }, updateObj, { returnOriginal: false });
+    eventEmitter.emit(CONSORTIUM_CHANGED, consortiaUpdateResult.value);
+  }
 }
 
 const pubsub = new PubSub();
@@ -165,44 +163,43 @@ const resolvers = {
      * Returns all results.
      * @return {array} All results
      */
-    fetchAllResults: () => fetchAll('runs'),
+    fetchAllResults: async () => {
+      const db = database.getDbInstance();
+
+      const results = await db.collection('runs').find().toArray();
+      return transformToClient(results);
+    },
     /**
      * Returns single pipeline
      * @param {object} args
      * @param {string} args.resultId  Requested pipeline ID
      * @return {object} Requested pipeline if id present, null otherwise
      */
-    fetchResult: (_, args) => {
+    fetchResult: async (_, args) => {
       if (!args.resultId) {
         return null;
-      } else {
-        return helperFunctions.getRethinkConnection()
-          .then(connection =>
-            rethink.table('runs')
-              .get(args.resultId)
-              .run(connection).then(res => connection.close().then(() => res))
-          )
-          .then(result => result);
       }
+
+      const db = database.getDbInstance();
+
+      const result = await db.collection('runs').findOne({ _id: ObjectID(args.resultId) });
+      return transformToClient(result);
     },
     /**
      * Fetches all public consortia and private consortia for which the current user has access
      * @return {array} All consortia to which the current user access
      */
     fetchAllConsortia: async ({ auth: { credentials } }) => {
-      const connection = await helperFunctions.getRethinkConnection();
+      const db = database.getDbInstance();
 
-      const cursor = await rethink
-        .table('consortia')
-        .orderBy({ index: 'id' })
-        .filter(rethink.row('isPrivate').eq(false).or(rethink.row('members').contains(credentials.id)))
-        .run(connection);
+      const consortia = await db.collection('consortia').find({
+        $or: [
+          { isPrivate: false },
+          { members: credentials.username }
+        ]
+      }).toArray();
 
-      const results = await cursor.toArray();
-
-      await connection.close();
-
-      return results;
+      return transformToClient(consortia);
     },
     /**
      * Returns single consortium.
@@ -210,55 +207,88 @@ const resolvers = {
      * @param {string} args.consortiumId Requested consortium ID
      * @return {object} Requested consortium if id present, null otherwise
      */
-    fetchConsortium: (_, args) => args.consortiumId ? fetchOne('consortia', args.consortiumId) : null,
+    fetchConsortium: async (_, args) => {
+      if (!args.consortiumId) {
+        return null;
+      }
+
+      const db = database.getDbInstance();
+
+      const consortium = await db.collection('consortia').findOne({ _id: ObjectID(args.consortiumId) });
+      return transformToClient(consortium);
+    },
     /**
      * Returns all computations.
      * @return {array} All computations
      */
-    fetchAllComputations: () => fetchAll('computations'),
+    fetchAllComputations: async () => {
+      const db = database.getDbInstance();
+
+      const computations = await db.collection('computations').find().toArray();
+      return transformToClient(computations);
+    },
     /**
      * Returns metadata for specific computation name
      * @param {object} args
      * @param {array} args.computationIds Requested computation ids
      * @return {array} List of computation objects
      */
-    fetchComputation: (_, args) => {
-      return helperFunctions.getRethinkConnection()
-        .then((connection) =>
-          rethink.table('computations').getAll(...args.computationIds)
-            .run(connection).then(res => connection.close().then(() => res))
-        )
-        .then((cursor) => cursor.toArray())
-        .then((result) => {
-          return result;
-        });
+    fetchComputation: async (_, { computationIds }) => {
+      if (!Array.isArray(computationIds) || computationIds.length === 0) {
+        return null;
+      }
+
+      const db = database.getDbInstance();
+
+      const computations = await db.collection('computations').find({
+        _id: { $in: computationIds.map(id => ObjectID(id)) }
+      }).toArray();
+
+      return transformToClient(computations);
     },
     /**
      * Returns all pipelines.
      * @return {array} List of all pipelines
      */
-    fetchAllPipelines: () => {
-      return helperFunctions.getRethinkConnection()
-        .then(connection =>
-          rethink.table('pipelines')
-            .orderBy({ index: 'id' })
-            .map(pipeline =>
-              pipeline.merge(pipeline =>
-                ({
-                  steps: pipeline('steps').map(step =>
-                    step.merge({
-                      computations: step('computations').map(compId =>
-                        rethink.table('computations').get(compId)
-                      )
-                    })
-                  )
-                })
-              )
-            )
-            .run(connection).then(res => connection.close().then(() => res))
-        )
-        .then(cursor => cursor.toArray())
-        .then(result => result);
+    fetchAllPipelines: async () => {
+      const db = database.getDbInstance();
+
+      const pipelineSteps = await db.collection('pipelines').aggregate([
+        { $unwind: '$steps' },
+        {
+          $lookup: {
+            from: 'computations',
+            localField: 'steps.computations',
+            foreignField: '_id',
+            as: 'steps.computations',
+          },
+        },
+      ]);
+
+      const pipelines = {};
+      while (await pipelineSteps.hasNext()) {
+        const currentStep = await pipelineSteps.next();
+
+
+        if (!(currentStep._id in pipelines)) {
+          pipelines[currentStep._id] = {
+            ...currentStep,
+            steps: [],
+          };
+        }
+
+        currentStep.steps.computations = transformToClient(currentStep.steps.computations);
+
+        pipelines[currentStep._id].steps.push(currentStep.steps);
+      }
+
+      const steplessPipelines = await db.collection('pipelines').find({
+        steps: { $size: 0 }
+      }).toArray();
+
+      steplessPipelines.forEach(p => pipelines[p._id] = p);
+
+      return transformToClient(Object.values(pipelines));
     },
     /**
      * Returns single pipeline
@@ -266,30 +296,13 @@ const resolvers = {
      * @param {string} args.pipelineId  Requested pipeline ID
      * @return {object} Requested pipeline if id present, null otherwise
      */
-    fetchPipeline: (_, args) => {
+    fetchPipeline: async (_, args) => {
       if (!args.pipelineId) {
         return null;
-      } else {
-        return helperFunctions.getRethinkConnection()
-          .then(connection =>
-            rethink.table('pipelines')
-              .get(args.pipelineId)
-              // Populate computations subfield with computation meta information
-              .merge(pipeline =>
-                ({
-                  steps: pipeline('steps').map(step =>
-                    step.merge({
-                      computations: step('computations').map(compId =>
-                        rethink.table('computations').get(compId)
-                      )
-                    })
-                  )
-                })
-              )
-              .run(connection).then(res => connection.close().then(() => res))
-          )
-          .then(result => result);
       }
+
+      const pipeline = await fetchOnePipeline(ObjectID(args.pipelineId));
+      return transformToClient(pipeline);
     },
     /**
      * Returns single user.
@@ -297,45 +310,41 @@ const resolvers = {
      * @param {string} args.userId Requested user ID, restricted to authenticated user for time being
      * @return {object} Requested user if id present, null otherwise
      */
-    fetchUser: ({ auth: { credentials } }, args) => {
-      if (args.userId !== credentials.id) {
+    fetchUser: async ({ auth: { credentials } }, args) => {
+      if (args.userId !== credentials.username) {
         return Boom.unauthorized('Unauthorized action');
       }
 
-      return fetchOne('users', credentials.id);
+      return helperFunctions.getUserDetails(credentials.username);
     },
-    fetchAllUsers: () => fetchAll('users'),
-    fetchAllUserRuns: ({ auth: { credentials } }, args) => {
-      let connection;
-      return helperFunctions.getRethinkConnection()
-        .then((db) => {
-          connection = db;
-          return rethink.table('runs')
-            .orderBy({ index: 'id' })
-            .filter(
-              rethink.row('clients').contains(credentials.id).
-              or(rethink.row('sharedUsers').contains(credentials.id))
-            )
-            .run(connection);
-          }
-        )
-        .then(cursor => cursor.toArray())
-        .then(res => connection.close().then(() => res));
+    fetchAllUsers: async () => {
+      const db = database.getDbInstance();
+
+      const users = await db.collection('users').find().toArray();
+      return transformToClient(users);
+    },
+    fetchAllUserRuns: async ({ auth: { credentials } }, args) => {
+      const db = database.getDbInstance();
+
+      const runs = await db.collection('runs').find({
+        $or: [
+          { clients: credentials.username },
+          { sharedUsers: credentials.username }
+        ]
+      }).toArray();
+
+      return transformToClient(runs);
     },
     fetchAllThreads: async ({ auth: { credentials } }) => {
-      const connection = await helperFunctions.getRethinkConnection();
-      const cursor = await rethink.table('threads')
-        .filter(doc => {
-          return doc('users').contains(user => {
-            return user('username').eq(credentials.id)
-          })
-        })
-        .run(connection)
-      const threads = cursor.toArray()
+      const db = database.getDbInstance();
 
-      await connection.close()
+      const threads = await db.collection('threads').find({
+        users: {
+          $elemMatch: { username: credentials.username }
+        }
+      }).toArray();
 
-      return threads
+      return transformToClient(threads);
     },
     validateComputation: (_, args) => {
       return new Promise();
@@ -343,26 +352,29 @@ const resolvers = {
   },
   Mutation: {
     /**
-     * Add computation to RethinkDB
+     * Add computation to database
      * @param {object} args
      * @param {object} args.computationSchema Computation object to add/update
      * @return {object} New/updated computation object
      */
-    addComputation: ({ auth: { credentials } }, args) => {
-      return helperFunctions.getRethinkConnection()
-        .then((connection) =>
-          rethink.table('computations').insert(
-            Object.assign({}, args.computationSchema, { submittedBy: credentials.id }),
-            {
-              conflict: "replace",
-              returnChanges: true,
-            }
-          )
-          .run(connection).then(res => connection.close().then(() => res))
-        )
-        .then((result) => {
-          return result.changes[0].new_val;
-        })
+    addComputation: async ({ auth: { credentials } }, args) => {
+      const db = database.getDbInstance();
+
+      args.computationSchema.id = args.computationSchema.id ? ObjectID(args.computationSchema.id) : new ObjectID();
+
+      await db.collection('computations').replaceOne({
+        _id: args.computationSchema.id,
+      }, {
+        ...args.computationSchema, submittedBy: credentials.id
+      }, {
+        upsert: true
+      });
+
+      const computation = transformToClient(args.computationSchema);
+
+      eventEmitter.emit(COMPUTATION_CHANGED, computation);
+
+      return computation;
     },
     /**
      * Add new user role to user perms, currently consortia perms only
@@ -381,57 +393,49 @@ const resolvers = {
         return Boom.forbidden('Action not permitted');
       }
 
-      const connection = await helperFunctions.getRethinkConnection();
-      await addUserPermissions(connection, args).then(res => connection.close().then(() => res));
+      await addUserPermissions({ doc: ObjectID(args.doc), role: args.role, userId: args.userId, table: args.table });
 
-      return helperFunctions.getUserDetails({ username: args.userId });
+      return helperFunctions.getUserDetails(args.userId);
     },
     /**
-     * Add run to RethinkDB
+     * Add run to database
      * @param {String} consortiumId Run object to add/update
      * @return {object} New/updated run object
      */
-    createRun: ({ auth }, { consortiumId }) => {
+    createRun: async ({ auth }, { consortiumId }) => {
       if (!auth || !auth.credentials) {
         // No authorized user, reject
         return Boom.unauthorized('User not authenticated');
       }
 
-      return fetchOne('consortia', consortiumId)
-        .then(consortium => Promise.all([
-          consortium,
-          fetchOnePipeline(consortium.activePipelineId),
-          helperFunctions.getRethinkConnection()
-        ]))
-        .then(([consortium, pipelineSnapshot, connection]) => {
-          const isPipelineDecentralized = pipelineSnapshot.steps.findIndex(step => step.controller.type === 'decentralized') > -1;
+      try {
+        const db = database.getDbInstance();
 
-          return rethink.table('runs').insert(
-            {
-              clients: [...consortium.members],
-              consortiumId,
-              pipelineSnapshot,
-              startDate: Date.now(),
-              type: isPipelineDecentralized ? 'decentralized' : 'local',
-            },
-            {
-              conflict: "replace",
-              returnChanges: true,
-            }
-          )
-          .run(connection).then(res => connection.close().then(() => res));
-        }
-        )
-        .then((result) => {
-          return axios.post(
-            `http://${config.host}:${config.pipelineServer}/startPipeline`, { run: result.changes[0].new_val }
-          ).then(() => {
-              return result.changes[0].new_val;
-          })
-        })
-        .catch(error => {
-              console.log(error)
+        const consortium = await db.collection('consortia').findOne({ _id: ObjectID(consortiumId) });
+        const pipeline = await fetchOnePipeline(consortium.activePipelineId);
+
+        const isPipelineDecentralized = pipeline.steps.findIndex(step => step.controller.type === 'decentralized') > -1;
+
+        const result = await db.collection('runs').insertOne({
+            clients: [...consortium.members],
+            consortiumId,
+            pipelineSnapshot: pipeline,
+            startDate: Date.now(),
+            type: isPipelineDecentralized ? 'decentralized' : 'local',
         });
+
+        const run = transformToClient(result.ops[0]);
+
+        await axios.post(
+          `http://${process.env.PIPELINE_SERVER_HOSTNAME}:${process.env.PIPELINE_SERVER_PORT}/startPipeline`, { run }
+        );
+
+        eventEmitter.emit(RUN_CHANGED, run);
+
+        return run;
+      } catch (error) {
+        console.log(error)
+      }
     },
     /**
      * Deletes consortium
@@ -440,29 +444,47 @@ const resolvers = {
      * @param {string} args.consortiumId Consortium id to delete
      * @return {object} Deleted consortium
      */
-    deleteConsortiumById: ({ auth: { credentials: { permissions } } }, args) => {
+    deleteConsortiumById: async ({ auth: { credentials: { permissions } } }, args) => {
       if (!permissions.consortia[args.consortiumId] || !permissions.consortia[args.consortiumId].includes('owner')) {
         return Boom.forbidden('Action not permitted');
       }
 
-      return helperFunctions.getRethinkConnection()
-        .then(connection =>
-          new Promise.all([
-            rethink.table('consortia').get(args.consortiumId)
-              .delete({ returnChanges: true })
-              .run(connection),
-            rethink.table('users').replace(user =>
-              user.without({
-                permissions: { consortia: args.consortiumId },
-                consortiaStatuses: args.consortiumId
-              })
-            ).run(connection),
-            rethink.table('pipelines').filter({ owningConsortium: args.consortiumId })
-              .delete()
-              .run(connection)
-          ]).then(res => connection.close().then(() => res))
-        )
-        .then(([consortium]) => consortium.changes[0].old_val)
+      const db = database.getDbInstance();
+
+      const deleteConsortiumResult = await db.collection('consortia').findOneAndDelete({ _id: ObjectID(args.consortiumId) });
+
+      eventEmitter.emit(CONSORTIUM_DELETED, deleteConsortiumResult.value);
+
+      const userIds = await db.collection('users').find({
+        [`permissions.consortia.${args.consortiumId}`]: { $exists: true }
+      }, {
+        projection: { _id: 1 }
+      }).toArray();
+
+      await db.collection('users').updateMany({
+        [`permissions.consortia.${args.consortiumId}`]: { $exists: true }
+      }, {
+        $unset: {
+          [`permissions.consortia.${args.consortiumId}`]: '',
+          [`consortiaStatuses.${args.consortiumId}`]: ''
+        }
+      });
+
+      const users = await db.collection('users').find({ _id: { $in: userIds.map(u => u._id) } }).toArray();
+
+      eventEmitter.emit(USER_CHANGED, users);
+
+      const pipelines = await db.collection('pipelines').find({
+        owningConsortium: ObjectID(args.consortiumId)
+      }).toArray();
+
+      await db.collection('pipelines').deleteMany({
+        owningConsortium: ObjectID(args.consortiumId)
+      });
+
+      eventEmitter.emit(PIPELINE_DELETED, pipelines);
+
+      return transformToClient(deleteConsortiumResult.value);
     },
     /**
      * Deletes pipeline
@@ -472,10 +494,11 @@ const resolvers = {
      * @return {object} Deleted pipeline
      */
     deletePipeline: async ({ auth: { credentials: { permissions } } }, args) => {
-      const connection = await helperFunctions.getRethinkConnection()
-      const pipeline = await rethink.table('pipelines')
-        .get(args.pipelineId)
-        .run(connection)
+      const db = database.getDbInstance();
+
+      const pipelineId = ObjectID(args.pipelineId);
+
+      const pipeline = await db.collection('pipelines').findOne({ _id: pipelineId });
 
       if (!permissions.consortia[pipeline.owningConsortium] ||
           !permissions.consortia[pipeline.owningConsortium].includes('owner')
@@ -483,28 +506,28 @@ const resolvers = {
         return Boom.forbidden('Action not permitted')
       }
 
-      const runsCount = await rethink.table('runs')('pipelineSnapshot')
-        .filter({ id: args.pipelineId })
-        .count()
-        .run(connection)
+      const runsCount = await db.collection('runs').countDocuments({
+        'pipelineSnapshot.id': args.pipelineId
+      });
 
       if (runsCount > 0) {
         return Boom.badData('Runs on this pipeline exist')
       }
 
-      await rethink.table('pipelines')
-        .get(args.pipelineId)
-        .delete({ returnChanges: true })
-        .run(connection)
+      const deletePipelineResult = await db.collection('pipelines').findOneAndDelete({ _id: pipelineId });
+      eventEmitter.emit(PIPELINE_DELETED, deletePipelineResult.value);
 
-      await rethink.table('consortia')
-        .filter({ activePipelineId: args.pipelineId })
-        .replace(rethink.row.without('activePipelineId'))
-        .run(connection)
+      const updateConsortiumResult = await db.collection('consortia').findOneAndUpdate({
+        activePipelineId: args.pipelineId
+      }, {
+        $unset: { activePipelineId: '' }
+      }, {
+        returnOriginal: false
+      });
 
-      await connection.close()
+      eventEmitter.emit(CONSORTIUM_CHANGED, updateConsortiumResult.value);
 
-      return pipeline
+      return transformToClient(deletePipelineResult.value);
     },
     /**
      * Add logged user to consortium members list
@@ -514,17 +537,16 @@ const resolvers = {
      * @return {object} Updated consortium
      */
     joinConsortium: async ({ auth: { credentials } }, args) => {
-      const connection = await helperFunctions.getRethinkConnection()
-      const consortium = await fetchOne('consortia', args.consortiumId)
+      const db = database.getDbInstance();
+      let consortium = await db.collection('consortia').findOne({ _id: ObjectID(args.consortiumId) });
 
-      if (consortium.members.indexOf(credentials.id) !== -1) {
-        return consortium
+      if (consortium.members.indexOf(credentials.username) !== -1) {
+        return consortium;
       }
 
-      await addUserPermissions(connection, { userId: credentials.id, role: 'member', doc: args.consortiumId, table: 'consortia' })
-      await connection.close()
+      await addUserPermissions({ userId: credentials.id, role: 'member', doc: ObjectID(args.consortiumId), table: 'consortia' });
 
-      return fetchOne('consortia', args.consortiumId)
+      return helperFunctions.getUserDetails(credentials.username);
     },
     /**
      * Remove logged user from consortium members list
@@ -534,11 +556,9 @@ const resolvers = {
      * @return {object} Updated consortium
      */
     leaveConsortium: async ({ auth: { credentials } }, args) => {
-      const connection = await helperFunctions.getRethinkConnection();
-      await removeUserPermissions(connection, { userId: credentials.id, role: 'member', doc: args.consortiumId, table: 'consortia' })
-        .then(res => connection.close().then(() => res));
+      await removeUserPermissions({ userId: credentials.id, role: 'member', doc: ObjectID(args.consortiumId), table: 'consortia' });
 
-      return fetchOne('consortia', args.consortiumId);
+      return helperFunctions.getUserDetails(credentials.username);
     },
     /**
      * Deletes computation
@@ -547,23 +567,20 @@ const resolvers = {
      * @param {string} args.computationId Computation id to delete
      * @return {object} Deleted computation
      */
-    removeComputation: ({ auth: { credentials } }, args) => {
-      return helperFunctions.getRethinkConnection()
-        .then((connection) =>
-          new Promise.all([
-            connection,
-            rethink.table('computations').get(args.computationId).run(connection)
-          ])
-        )
-        .then(([connection, comp]) => {
-          if (comp.submittedBy !== credentials.id) {
-            return Boom.forbidden('Action not permitted');
-          }
+    removeComputation: async ({ auth: { credentials } }, args) => {
+      const db = database.getDbInstance();
 
-          return rethink.table('computations').get(args.computationId)
-            .delete({ returnChanges: true }).run(connection).then(res => connection.close().then(() => res))
-        })
-        .then(result => result.changes[0].old_val)
+      const computation = await db.collection('computations').findOne({ _id: ObjectID(args.computationId) });
+
+      if (computation.submittedBy !== credentials.username) {
+        return Boom.forbidden('Action not permitted');
+      }
+
+      const deleteComputationResult = await db.collection('computations').findOneAndDelete({ _id: ObjectID(args.computationId) });
+
+      eventEmitter.emit(COMPUTATION_DELETED, deleteComputationResult.value);
+
+      return transformToClient(deleteComputationResult.value);
     },
     /**
      * Add new user role to user perms, currently consortia perms only
@@ -573,20 +590,18 @@ const resolvers = {
      * @param {string} args.table Table of the document to add role to
      * @param {string} args.doc Id of the document to add role to
      * @param {string} args.role Role to add to perms
-     * @param {string} args.userId Id of the user to be removed
      * @return {object} Updated user object
      */
     removeUserRole: async ({ auth: { credentials } }, args) => {
-      const { permissions } = credentials
+      const { permissions } = credentials;
 
       if (!permissions[args.table][args.doc] || !permissions[args.table][args.doc].includes('owner')) {
         return Boom.forbidden('Action not permitted');
       }
 
-      const connection = await helperFunctions.getRethinkConnection();
-      await removeUserPermissions(connection, args)
-        .then(res => connection.close().then(() => res));
-      return helperFunctions.getUserDetails({ username: args.userId });
+      await removeUserPermissions({ doc: ObjectID(args.doc), role: args.role, userId: args.userId, table: args.table });
+
+      return helperFunctions.getUserDetails(args.userId);
     },
     /**
      * Sets active pipeline on consortia object
@@ -604,14 +619,22 @@ const resolvers = {
             return Boom.forbidden('Action not permitted');
       }*/
 
-      const connection = await helperFunctions.getRethinkConnection()
-      const result = await rethink.table('consortia')
-        .get(args.consortiumId)
-        .update({ activePipelineId: args.activePipelineId, mappedForRun: [] })
-        .run(connection)
-      await connection.close()
+      const db = database.getDbInstance();
 
-      return result
+      const result = await db.collection('consortia').findOneAndUpdate({
+        _id: ObjectID(args.consortiumId)
+      }, {
+        $set: {
+          activePipelineId: ObjectID(args.activePipelineId),
+          mappedForRun: []
+        }
+      }, {
+        returnOriginal: false
+      });
+
+      eventEmitter.emit(CONSORTIUM_CHANGED, result.value);
+
+      return transformToClient(result.value);
     },
     /**
      * Saves consortium
@@ -629,43 +652,42 @@ const resolvers = {
         return Boom.forbidden('Action not permitted');
       }
 
-      const connection = await helperFunctions.getRethinkConnection();
+      const db = database.getDbInstance();
+
+      const consortiumData = Object.assign(
+        { ...args.consortium },
+        !isUpdate && { createDate: Date.now() },
+      );
 
       if (!isUpdate) {
-        const count = await rethink.table('consortia')
-          .filter({ name: args.consortium.name })
-          .count()
-          .run(connection)
+        const count = await db.collection('consortia').countDocuments({
+          name: args.consortium.name
+        });
 
         if (count > 0) {
           return Boom.forbidden('Consortium with same name already exists');
         }
       }
 
-      const consortiumData = Object.assign(
-        { ...args.consortium },
-        !isUpdate && { createDate: Date.now() },
-      )
+      consortiumData.id = consortiumData.id ? ObjectID(consortiumData.id) : new ObjectID();
 
-      const result = await rethink.table('consortia')
-        .insert(consortiumData, {
-          conflict: 'update',
-          returnChanges: true,
-        })
-        .run(connection);
-
-      const consortiumId = consortiumData.id || result.changes[0].new_val.id;
-
-      if (!isUpdate) {
-        await addUserPermissions(connection, { userId: credentials.id, role: 'owner', doc: consortiumId, table: 'consortia' });
-        await addUserPermissions(connection, { userId: credentials.id, role: 'member', doc: consortiumId, table: 'consortia' });
+      if (consortiumData.activePipelineId) {
+        consortiumData.activePipelineId = ObjectID(consortiumData.activePipelineId);
       }
 
-      const consortium = await fetchOne('consortia', consortiumId);
+      await db.collection('consortia').replaceOne({
+        _id: consortiumData.id
+      }, consortiumData, {
+        upsert: true,
+      });
 
-      await connection.close();
+      if (!isUpdate) {
+        await addUserPermissions({ userId: credentials.id, role: 'owner', doc: consortiumData.id, table: 'consortia' });
+        await addUserPermissions({ userId: credentials.id, role: 'member', doc: consortiumData.id, table: 'consortia' });
+      }
 
-      return consortium;
+      const consortium = await db.collection('consortia').findOne({ _id: consortiumData.id });
+      return transformToClient(consortium);
     },
     /**
      * Saves run error
@@ -674,10 +696,14 @@ const resolvers = {
      * @param {string} args.runId Run id to update
      * @param {string} args.error Error
      */
-    saveError: async ({ auth: { credentials } }, args) => {
-      const connection = await helperFunctions.getRethinkConnection();
+    saveError: async (_, args) => {
+      const db = database.getDbInstance();
 
-      const run = await rethink.table('runs').get(args.runId).run(connection);
+      const run = await db.collection('runs').findOne({
+        _id: ObjectID(args.runId)
+      }, {
+        projection: { _id: 1, type: 1 }
+      });
 
       if (!run) {
         return;
@@ -691,9 +717,17 @@ const resolvers = {
         updateObj.error = Object.assign({}, args.error);
       }
 
-      await rethink.table('runs').get(args.runId).update(updateObj).run(connection);
+      const result = await db.collection('runs').findOneAndUpdate({
+        _id: ObjectID(args.runId)
+      }, {
+        $set: updateObj
+      }, {
+        returnOriginal: false
+      });
 
-      return connection.close();
+      eventEmitter.emit(RUN_CHANGED, result.value);
+
+      return transformToClient(result.value);
     },
     /**
      * Saves pipeline
@@ -710,18 +744,7 @@ const resolvers = {
           && !permissions.consortia[args.consortium.id].write) {
             return Boom.forbidden('Action not permitted');
       }*/
-      const connection = await helperFunctions.getRethinkConnection();
-
-      if (!args.pipeline.id) {
-        const count = await rethink.table('pipelines')
-          .filter({ name: args.pipeline.name })
-          .count()
-          .run(connection)
-
-        if (count > 0) {
-          return Boom.forbidden('Pipeline with same name already exists');
-        }
-      }
+      const db = database.getDbInstance();
 
       if (args.pipeline && args.pipeline.steps) {
         const invalidData = args.pipeline.steps.some(step =>
@@ -738,17 +761,28 @@ const resolvers = {
         }
       }
 
-      const result = await rethink.table('pipelines')
-        .insert(args.pipeline, {
-          conflict: 'update',
-          returnChanges: true,
-        })
-        .run(connection);
+      args.pipeline.id = args.pipeline.id ? ObjectID(args.pipeline.id) : new ObjectID();
 
-      const pipelineId = args.pipeline.id || result.changes[0].new_val.id;
-      const pipeline = await fetchOnePipeline(pipelineId);
+      if (args.pipeline.owningConsortium) {
+        args.pipeline.owningConsortium = ObjectID(args.pipeline.owningConsortium);
+      }
 
-      await connection.close();
+      if (args.pipeline.steps) {
+        args.pipeline.steps.forEach(step => {
+          if (step.computations) {
+            step.computations = step.computations.map(compId => ObjectID(compId));
+          }
+        });
+      }
+
+      await db.collection('pipelines').replaceOne({
+        _id: args.pipeline.id
+      }, args.pipeline, {
+        upsert: true
+      });
+
+      const pipeline = await fetchOnePipeline(ObjectID(args.pipeline.id));
+      eventEmitter.emit(PIPELINE_CHANGED, pipeline);
 
       return pipeline;
     },
@@ -759,10 +793,14 @@ const resolvers = {
      * @param {string} args.runId Run id to update
      * @param {string} args.results Results
      */
-    saveResults: async ({ auth: { credentials } }, args) => {
-      const connection = await helperFunctions.getRethinkConnection();
+    saveResults: async (_, args) => {
+      const db = database.getDbInstance();
 
-      const run = await rethink.table('runs').get(args.runId).run(connection);
+      const run = await db.collection('runs').findOne({
+        _id: ObjectID(args.runId)
+      }, {
+        projection: { _id: 1, type: 1 }
+      });
 
       if (!run) {
         return;
@@ -776,9 +814,17 @@ const resolvers = {
         updateObj.results = Object.assign({}, args.results);
       }
 
-      await rethink.table('runs').get(args.runId).update(updateObj).run(connection);
+      const result = await db.collection('runs').findOneAndUpdate({
+        _id: ObjectID(args.runId)
+      }, {
+        $set: updateObj
+      }, {
+        returnOriginal: false
+      });
 
-      return connection.close();
+      eventEmitter.emit(RUN_CHANGED, result.value);
+
+      return transformToClient(result.value);
     },
     setActiveComputation: (_, args) => {
       return new Promise();
@@ -793,14 +839,22 @@ const resolvers = {
      * @param {string} args.runId Run id to update
      * @param {string} args.data State data
      */
-    updateRunState: ({ auth: { credentials } }, args) => {
-      const { permissions } = credentials;
-      return helperFunctions.getRethinkConnection()
-        .then((connection) => {
-          return rethink.table('runs').get(args.runId).update({ remotePipelineState: args.data })
-          .run(connection).then(res => connection.close());
-        });
-          // .then(result => result.changes[0].new_val)
+    updateRunState: async (_, args) => {
+      const db = database.getDbInstance();
+
+      const result = await db.collection('runs').findOneAndUpdate({
+        _id: ObjectID(args.runId)
+      }, {
+        $set: {
+          remotePipelineState: args.data,
+        }
+      }, {
+        returnOriginal: false
+      });
+
+      eventEmitter.emit(RUN_CHANGED, result.value);
+
+      return transformToClient(result.value);
     },
     /**
      * Saves consortium
@@ -810,21 +864,23 @@ const resolvers = {
      * @param {string} args.status New status
      * @return {object} Updated user object
      */
-    updateUserConsortiumStatus: ({ auth: { credentials } }, { consortiumId, status }) =>
-      helperFunctions.getRethinkConnection()
-        .then(connection =>
-          rethink.table('users')
-          .get(credentials.id).update({
-            consortiaStatuses: {
-              [consortiumId]: status,
-            },
-          }).run(connection).then(res => connection.close().then(() => res))
-        )
-        .then(result =>
-          helperFunctions.getUserDetails({ username: credentials.id })
-        )
-        .then(result => result)
-    ,
+    updateUserConsortiumStatus: async ({ auth: { credentials } }, { consortiumId, status }) => {
+      const db = database.getDbInstance();
+
+      const result = await db.collection('users').findOneAndUpdate({
+        _id: credentials.id
+      }, {
+        $set: {
+          [`consortiaStatuses.${consortiumId}`]: status
+        }
+      }, {
+        returnOriginal: false
+      });
+
+      eventEmitter.emit(USER_CHANGED, result.value);
+
+      return transformToClient(result.value);
+    },
     /**
      * Updated consortium mapped users
      * @param {object} auth User object from JWT middleware validateFunc
@@ -834,13 +890,21 @@ const resolvers = {
      * @return {object} Updated consortia
      */
     updateConsortiumMappedUsers: async ({ auth: { credentials } }, args) => {
-      const connection = await helperFunctions.getRethinkConnection()
-      const result = await rethink.table('consortia')
-        .get(args.consortiumId)
-        .update({ mappedForRun: args.mappedForRun })
-        .run(connection)
-      await connection.close()
-      return result
+      const db = database.getDbInstance();
+
+      const result = await db.collection('consortia').findOneAndUpdate({
+        _id: ObjectID(args.consortiumId)
+      }, {
+        $set: {
+          mappedForRun: args.mappedForRun
+        }
+      }, {
+        returnOriginal: false
+      });
+
+      eventEmitter.emit(CONSORTIUM_CHANGED, result.value);
+
+      return transformToClient(result.value);
     },
     /**
      * Updated consortia mapped users
@@ -850,14 +914,35 @@ const resolvers = {
      * @return {object} Updated consortia
      */
     updateConsortiaMappedUsers: async ({ auth: { credentials } }, args) => {
-      const connection = await helperFunctions.getRethinkConnection()
-      const result = await rethink.table('consortia')
-        .getAll(...args.consortia)
-        .filter(rethink.row('mappedForRun').contains(credentials.id))
-        .update({ mappedForRun: rethink.row('mappedForRun').difference([credentials.id]) })
-        .run(connection)
+      if (!Array.isArray(args.consortia) || args.consortia.length === 0) {
+        return;
+      }
 
-      return result
+      const db = database.getDbInstance();
+
+      const consortiaIds = args.consortia.map(id => ObjectID(id));
+
+      const updatedConsortiaIds = await db.collection('consortia').find({
+        _id: { $in: consortiaIds },
+        mappedForRun: credentials.username
+      }, {
+        projection: { _id: 1 }
+      }).toArray();
+
+      await db.collection('consortia').updateMany({
+        _id: { $in: consortiaIds },
+        mappedForRun: credentials.username
+      }, {
+        $pull: {
+          mappedForRun: credentials.username
+        }
+      });
+
+      const consortia = await db.collection('consortia').find({
+        _id: { $in: updatedConsortiaIds.map(c => c._id) }
+      });
+
+      eventEmitter.emit(CONSORTIUM_CHANGED, consortia);
     },
     /**
      * Save message
@@ -871,72 +956,73 @@ const resolvers = {
      * @return {object} Updated message
      */
     saveMessage: async ({ auth: { credentials } }, args) => {
-      const { threadId, title, recipients, content, action } = args
+      const { title, recipients, content, action } = args;
+      const threadId = ObjectID(args.threadId);
 
-      const connection = await helperFunctions.getRethinkConnection()
+      const db = database.getDbInstance();
 
       const messageToSave = Object.assign(
         {
-          id: rethink.uuid(),
+          _id: new ObjectID(),
           sender: credentials.id,
           recipients,
           content,
           date: Date.now(),
         },
         action && { action },
-      )
+      );
 
-      let result
+      let result;
 
       if (threadId) {
-        const thread = await fetchOne('threads', threadId)
-        const { messages, users } = thread
-        const threadToSave = {
-          messages: [...messages, messageToSave],
-          users: uniq([...users.map(user => user.username), ...recipients])
-            .map(user => ({ username: user, isRead: user === credentials.id })),
-          date: Date.now(),
-        }
+        const thread = await db.collection('threads').findOne({ _id: threadId });
 
-        await rethink.table('threads')
-          .get(threadId)
-          .update(threadToSave, { nonAtomic: true })
-          .run(connection)
+        const { users } = thread;
 
-        result = await fetchOne('threads', threadId)
+        const updateObj = {
+          $push: {
+            messages: messageToSave
+          },
+          $set: {
+            date: Date.now(),
+            users: uniq([...users.map(user => user.username), ...recipients])
+              .map(user => ({ username: user, isRead: user === credentials.id }))
+          }
+        };
+
+        const updateResult = await db.collection('threads').findOneAndUpdate({ _id: threadId }, updateObj, { returnOriginal: false });
+
+        result = updateResult.value;
       } else {
         const thread = {
-          id: rethink.uuid(),
           owner: credentials.id,
           title: title,
           messages: [messageToSave],
-          users: uniq([credentials.id, ...recipients])
-            .map(user => ({ username: user, isRead: user === credentials.id })),
+          users: uniq([credentials.username, ...recipients])
+            .map(user => ({ username: user, isRead: user === credentials.username })),
           date: Date.now(),
-        }
+        };
 
-        const res = await rethink.table('threads')
-          .insert(thread, { returnChanges: true })
-          .run(connection)
+        const insertResult = await db.collection('threads').insertOne(thread);
 
-        result = res.changes[0].new_val
+        result = insertResult.ops[0];
       }
 
       if (action && action.type === 'share-result') {
-        const run = await fetchOne('runs', action.detail.id)
-        const runToSave = {
-          sharedUsers: uniq([...run.sharedUsers || [], ...recipients]),
-        }
+        const updateRunResult = await db.collection('runs').findOneAndUpdate({ _id: ObjectID(action.detail.id) }, {
+          $addToSet: {
+            sharedUsers: { $each: recipients }
+          }
+        }, {
+          returnOriginal: false
+        });
 
-        await rethink.table('runs')
-          .get(action.detail.id)
-          .update(runToSave, { nonAtomic: true })
-          .run(connection)
+        eventEmitter.emit(RUN_CHANGED, updateRunResult.value);
       }
 
-      await connection.close()
+      eventEmitter.emit(THREAD_CHANGED, result);
 
-      return result
+      return transformToClient(result);
     },
     /**
      * Set read mesasge
@@ -946,25 +1032,25 @@ const resolvers = {
      * @param {string} args.userId User Id
      * @return {object} None
      */
-    setReadMessage: async ({ auth: { credentials } }, args) => {
-      const { threadId, userId } = args
+    setReadMessage: async (_, args) => {
+      const { threadId, userId } = args;
 
-      const connection = await helperFunctions.getRethinkConnection()
-      const thread = await fetchOne('threads', threadId)
+      const db = database.getDbInstance();
 
-      const threadToSave = {
-        ...thread,
-        users: thread.users.map(user => user.username === userId ? { ...user, isRead: true } : user)
-      }
+      const result = await db.collection('threads').findOneAndUpdate({
+        _id: ObjectID(threadId),
+        'users.username': userId,
+      }, {
+        $set: {
+          'users.$.isRead': true,
+        }
+      }, {
+        returnOriginal: false
+      });
 
-      await rethink.table('threads')
-        .get(threadId)
-        .update(threadToSave, { nonAtomic: true })
-        .run(connection)
+      eventEmitter.emit(THREAD_CHANGED, result.value);
 
-      await connection.close()
-
-      return
+      return transformToClient(result.value);
     }
   },
   Subscription: {
@@ -1030,20 +1116,7 @@ const resolvers = {
     userChanged: {
       subscribe: withFilter(
         () => pubsub.asyncIterator('userChanged'),
-        (payload, variables) => (variables.userId || payload.userId === variables.userId)
-      )
-    },
-    /**
-     * User Metadata subscription
-     * @param {object} payload
-     * @param {string} payload.userId The user changed
-     * @param {object} variables
-     * @param {string} variables.userId The user listened for
-     */
-    userMetadataChanged: {
-      subscribe: withFilter(
-        () => pubsub.asyncIterator('userMetadataChanged'),
-        (payload, variables) => (variables.userId && payload.userId === variables.userId)
+        (payload, variables) => (!variables.userId || payload.userId === variables.userId)
       )
     },
     /**
