@@ -1,4 +1,5 @@
 'use strict';
+const _ = require('lodash');
 
 module.exports = {
   create(spec, mode, runId, clientId, docker) {
@@ -8,7 +9,8 @@ module.exports = {
       mode === 'remote' ? spec.computation.remote : {}
     );
     const { meta } = spec;
-
+    const dockerOptions = mode === 'remote'
+      ? (spec.computation.remote.dockerOptions || {}) : (spec.computation.dockerOptions || {});
     return {
       computation,
       meta,
@@ -21,7 +23,7 @@ module.exports = {
           this.meta.id,
           `${this.runId}-${this.clientId}`,
           {
-            docker: {
+            docker: _.merge({
               Image: computation.dockerImage,
               HostConfig: {
                 Binds: [
@@ -31,24 +33,11 @@ module.exports = {
                   `${baseDirectory}/transfer:/transfer:rw`,
                 ],
               },
-            },
+            }, dockerOptions),
           }
         )
           .then((service) => {
-            return service(computation.command.concat([`${JSON.stringify(input)}`]))
-              .catch((error) => {
-                // decorate w/ input
-                error.input = input;
-                throw error;
-              })
-              .then((data) => {
-                if (typeof data === 'string') {
-                  const err = new Error(`Computation output serialization failed with value: ${data}`);
-                  err.input = input;
-                  throw err;
-                }
-                return data;
-              });
+            return service(computation.command.concat([`${JSON.stringify(input)}`]));
           });
       },
       stop() {
