@@ -8,7 +8,7 @@ import { withStyles } from '@material-ui/core/styles';
 const styles = theme => ({
   containerStyles: {
     display: 'inline-block',
-    margin: theme.spacing.unit,
+    margin: theme.spacing(1),
     verticalAlign: 'top',
     textAlign: 'center',
     position: 'relative',
