@@ -30,7 +30,7 @@ const styles = theme => ({
     overflow: 'hidden',
   },
   chip: {
-    margin: `${theme.spacing.unit / 2}px ${theme.spacing.unit / 4}px`,
+    margin: `${theme.spacing(1) / 2}px ${theme.spacing(1) / 4}px`,
   },
   chipFocused: {
     backgroundColor: emphasize(
@@ -44,7 +44,7 @@ const styles = theme => ({
     color: 'white',
   },
   noOptionsMessage: {
-    padding: `${theme.spacing.unit}px ${theme.spacing.unit * 2}px`,
+    padding: `${theme.spacing(1)}px ${theme.spacing(1) * 2}px`,
   },
   singleValue: {
     fontSize: 16,
@@ -57,12 +57,12 @@ const styles = theme => ({
   paper: {
     position: 'absolute',
     zIndex: 1,
-    marginTop: theme.spacing.unit,
+    marginTop: theme.spacing(1),
     left: 0,
     right: 0,
   },
   divider: {
-    height: theme.spacing.unit * 2,
+    height: theme.spacing(2),
   },
 });
 
