@@ -18,25 +18,25 @@ import TimeAgo from './time-ago';
 const styles = theme => ({
   rootPaper: {
     ...theme.mixins.gutters(),
-    paddingTop: theme.spacing.unit * 2,
-    paddingBottom: theme.spacing.unit * 2,
-    marginTop: theme.spacing.unit * 2,
+    paddingTop: theme.spacing(2),
+    paddingBottom: theme.spacing(2),
+    marginTop: theme.spacing(2),
   },
   titleContainer: {
     display: 'flex',
     justifyContent: 'space-between',
-    marginBottom: theme.spacing.unit,
+    marginBottom: theme.spacing(1),
   },
   label: {
     display: 'inline-block',
     fontWeight: 'bold',
-    marginRight: theme.spacing.unit,
+    marginRight: theme.spacing(1),
   },
   value: {
     display: 'inline-block',
   },
   contentContainer: {
-    marginBottom: theme.spacing.unit,
+    marginBottom: theme.spacing(1),
   },
   actionButtons: {
     display: 'flex',
@@ -50,11 +50,11 @@ const styles = theme => ({
     justifyContent: 'space-between',
   },
   runStatePaper: {
-    width: `calc(50% - ${theme.spacing.unit}px)`,
+    width: `calc(50% - ${theme.spacing(1)}px)`,
   },
   runTitle: {
     textDecoration: 'underline',
-    marginBottom: theme.spacing.unit,
+    marginBottom: theme.spacing(1),
   },
 });
 
@@ -67,7 +67,7 @@ function getStateWell(runObject, stateName, stateKey, classes) {
     <Paper
       className={classNames(classes.rootPaper, classes.runStatePaper)}
     >
-      <Typography variant="headline" className={classes.runTitle}>
+      <Typography variant="h5" className={classes.runTitle}>
         {`${stateName} Pipeline State:`}
       </Typography>
       {
@@ -156,7 +156,7 @@ class RunItem extends Component {
         className={classNames(classes.rootPaper, 'run-item-paper')}
       >
         <div className={classes.titleContainer}>
-          <Typography variant="headline">
+          <Typography variant="h5">
             { consortiumName }
             {
               pipelineSnapshot
@@ -166,7 +166,7 @@ class RunItem extends Component {
           {
             !endDate && status === 'started'
             && (
-              <Typography variant="headline">
+              <Typography variant="h5">
                 {'Started: '}
                 <TimeAgo timestamp={runObject.startDate / 1000} />
               </Typography>
@@ -175,7 +175,7 @@ class RunItem extends Component {
           {
             endDate
             && (
-              <Typography variant="headline">
+              <Typography variant="h5">
                 {'Completed: '}
                 <TimeAgo timestamp={runObject.endDate / 1000} />
               </Typography>
