@@ -15,11 +15,11 @@ import { notifySuccess } from '../../state/ducks/notifyAndLog';
 
 const styles = theme => ({
   topMargin: {
-    marginTop: theme.spacing.unit,
+    marginTop: theme.spacing(1),
   },
   description: {
-    marginTop: theme.spacing.unit * 2,
-    marginBottom: theme.spacing.unit * 2,
+    marginTop: theme.spacing(2),
+    marginBottom: theme.spacing(2),
   },
   actionsContainer: {
     display: 'flex',
@@ -82,7 +82,7 @@ class ComputationSubmission extends Component { // eslint-disable-line
             Computation Submission:
           </Typography>
         </div>
-        <Typography variant="body1" className={classes.description}>
+        <Typography variant="body2" className={classes.description}>
           Use the button below to upload your schema for review. Prior to submission,
           your schema will be validated. Please fix any errors found therein to unlock the
           <span style={{ fontWeight: 'bold' }}> Submit </span>
