@@ -10,12 +10,12 @@ import { pipelineNeedsDataMapping } from '../../../main/utils/run-pipeline-funct
 
 const styles = theme => ({
   contentContainer: {
-    marginTop: theme.spacing.unit,
-    marginBottom: theme.spacing.unit,
+    marginTop: theme.spacing(1),
+    marginBottom: theme.spacing(1),
   },
   labelInline: {
     fontWeight: 'bold',
-    marginRight: theme.spacing.unit,
+    marginRight: theme.spacing(1),
     display: 'inline-block',
   },
   value: {
@@ -106,7 +106,7 @@ class MapsList extends Component {
         </div>
         <Grid
           container
-          spacing={16}
+          spacing={2}
           direction="row"
           alignItems="stretch"
         >
