@@ -45,7 +45,7 @@ const styles = theme => ({
     padding: 15,
   },
   title: {
-    marginBottom: theme.spacing.unit * 2,
+    marginBottom: theme.spacing(2),
   },
   container: {
     flex: 1,
